CR feedback: TrendlineOption.tsx as separate file, remove TRENDLINE_VALUES array, remove usages of any type, const/let
